+++ lib/src/move/move_meta.dart
@@ -1,0 +1,10 @@
+part of 'move.dart';
+
+class MoveMeta {
+  final String algebraic;
+  final String prettyName;
+  const MoveMeta({
+    required this.algebraic,
+    required this.prettyName,
+  });
+}